Fix to RainBrowser to deal with this issue:
http://rainmeter.net/forum/viewtopic.php?p=10952#p10952

Still hoping we can enable !RainmeterActivateConfig and !RainmeterQuit when no configs are active, but this will work for now.  

Basically if there are no configs loaded and you ask to load one, I set "Active=n" (whatever is appropriate for the .ini in question) in Rainmeter.ini if the config exists there, or add [ConfigName] and "Active=n" if not, then kill and start Rainmeter using ProcessClose() and ShellExecute().