New !bang
!RainmeterRefreshApp

Does a full refresh of all skins and reloads the list of configs and Rainmeter.ini settings.  This is the same as "Refresh All" from the system tray context menu.  Is not the same as !RainmeterRefresh [config | *] as it can be used to make Rainmeter "see" a new folder added to \Skins by the user or a process.  Note: This was added mostly to better support RainBrowser when a user adds a downloaded config while using RainBrowser.