Rainstaller:
This commit is to get the repository set up to add Rainstaller.exe to the Rainmeter build and installation package

- Rainstaller is installed to the Rainmeter program folder under \Addons\Rainstaller
Rainstaller.exe
Rainstaller.chm

- A file association is created for the extension .rmskins to allow double clicking a .rmskins file to open Rainstaller.exe, unpack the .rmskins archive (it is a renamed .zip file) and install the Skin(s) / Themes included.

- Rainstaller.lnk (Rainstaller.exe) and Rainstaller Help.lnk (Rainstaller Help.chm) are added to the Windows Start Menu under Rainmeter

- Code added to the "uninstall" routines in Rainmeter Install Script.nsi to remove the file association for Rainstaller.exe / .rmskins on uninstall of Rainmeter and to remove the links from the Windows Start Menu.

*****************
Code Changes:

- New version of Build.py for compiling all source code
- New version of Rainmeter Install Script.nsi for creating the Rainmeter installation package

- Source code and build script for Rainstaller in \Addons\Rainstaller\Source
Rainmeter.nsi
Rainmeter.chm
\include (folder has some addon scripts and plugins for NSIS)

Note:  As of this build, everything works EXCEPT that Rainstaller can only be used by double-clicking a .rmskins file.  The ability to drag and drop a .rmskins file on Rainstaller.exe or to run Rainstaller.exe and "browse" to open a .rmskins file is coming in the next day or two.