Update Histogram meter to Direct2D (For real this time)
There are still quite a few things to do with this, but we need to
update bitmaps to be Direct2D instead of gdi+.

Fair warning image stuff is totally broken with this meter currently as
i can't use a Direct2D clip with a gdi bitmap. The logic is there, the
bitmaps are not though :)

We should probably change the "bothpath" to be a xor combination of the
parimary path and the secondary path, it should be a bit better.