Added unique random support to text files
Rewrote how the strings are selected so that every string in the file
has an equal likelyhood of being selected instead of heavily favoring a
the next string after having not had a seperator for a while. This is
done by preparsing the file similar to how directories are done so we
know where each different string starts and ends. With these changes
adding support for UniqueRandom is trivial.

Also will reread the file when the seperator is not where it is expected
to be. If this happens then it is safe to assume that the file has been
modified and needs reread. This maintains BWC for skins that modify text
files on the file that are used in quote meter.