UsageMonitor bug fixes
Sorry this is all in one commit, my individual commits were borked.

Alias=GPU now has a MaxValue of 100, this and measures with Percent=1
are the only ones that will have a max value by default since the others
are all dynamic.
Handled crashes when updating counters will now affect only that counter
or instance as opposed to always causing the ramining counters in that
category to fail.
Percent=1 on counters that don't have _Total now just return the normal
value.
Fixed a crash with measures that had PID translation turned on such as
Alias=GPU.