MeasureProcess: Remove use of background thread
Previously, a background thread woke up to update the process list 250ms. This
can be reduced to a function that caches the results and updates them if 250ms
has elapsed since the last call.

We also now store the process names lowercased in a std::unordered_set to
avoid iterating over a std::vector.

Both of these optimizations should reduce the amount of work done.