Add GitHub Actions job for builds (#250)
We now use CI for official, codesigned builds. This also updates our
codesigning certificate to a fresh version that expires in Jan 2024.

This also makes some changes to how Build.bat is supposed to be used. In
particular, the version is no longer hardcoded in neither Build.bat nor
Version.h/Version.cs and instead should be passed as parameters to
Build.bat. See updates to Building.md for more info.

As part of this, the revisions are no longer tied to the number of
commits. This was fine with SVN, but has never really worked well with
Git's merge commits (e.g. 990dee08e9da349a576b0308a13f51ba446e67eb). The
revision is now an arbitrary number just like the other version
components. Additionally, the version of local dev builds is now
`0.0.0.0`.