ActionTimer: Fix warning from 244143e
This needs to account for a 0 or negative "Repeat", so int should be used.

Honestly, I think this might a compiler bug. "_wtoi" returns an int, "2" is an int, and "1" is an int. So you get "(int * int) - int", so there should be no casting to another type. Removing the parenthesis removed the warning.