Fix NowPlaying for Spotify Premium by making it scan the windowTitle for "Spotify"
By default, NowPlaying in Rainmeter looks at the length of the window title, and makes sure it is longer than 7 characters (the length of "Spotify"), in order to detect playing media. When using Spotify Premium, it is default longer than 7 characters, so it thinks a song is always playing, and does not return the correct value for the state of the player. By scanning for the string "Spotify" in the window title, you can determine whether the media is paused, regardless of whether you use free or premium