Convert font enumeration from Gdiplus to DirectWrite
Note: This is just for logging purposes and does not change any font loading for the String meter. In general, this will log more font families since font faces will now be included in the list.

Also: In debug mode, this will also enumerate any fonts loaded with LocalFont and the fonts in "@Resources\Fonts" folder when a skin loads. This will help skin authors find the correct font name to use with the FontFace option.