Remove non-obsolete languages from status.json
This reduces the bandwidth for downloading status.json since we the code
only cares about obsolete languages. We additionally remove the language
name from status.json since it was not used in code. To make sense of
the numbers for humans, we add a languages.txt with a mapping from the
number to the name.